Add a file for scratch-blocks-specific util functions
